# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Todolist API
  version: 1.2.0 # Incremented version
  description: |
    API for managing Todo items, including CRUD operations, subtasks, deadlines, attachments, and user-defined Tags.
    Supports user authentication via email/password (JWT) and Google OAuth.
    Designed for use with oapi-codegen and Chi.

    **Note on Notifications:** Real-time notifications (e.g., via SSE or WebSockets) are planned but not fully described in this OpenAPI specification due to limitations in representing asynchronous APIs. These will be documented separately.

    **Note on Tag Deletion:** Deleting a Tag will typically remove its association from any Todo items currently using it.
servers:
  # The base path for all API routes defined below.
  # oapi-codegen will use this when setting up routes with HandlerFromMux.
  - url: /api/v1
    description: API version 1
components:
...
  securitySchemes:
    BearerAuth: # Used by API clients (non-browser)
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token provided in the Authorization header.
    CookieAuth: # Used by the web application (browser)
      type: apiKey
      in: cookie
      name: jwt_token # Name needs to match config.AppConfig.CookieName
      description: JWT authentication token provided via an HTTP-only cookie.

  # Reusable Schemas
  schemas:
...
    User:
      type: object
      description: Represents a registered user.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        username:
          type: string
        email:
          type: string
          format: email
        emailVerified:
          type: boolean
          readOnly: true
          description: Indicates if the user's email has been verified (e.g., via OAuth or email confirmation).
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      required:
        - id
        - username
        - email
        - emailVerified
        - createdAt
        - updatedAt
    SignupRequest:
      type: object
      description: Data required for signing up a new user via email/password.
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
          writeOnly: true # Password should not appear in responses
      required:
        - username
        - email
        - password
    LoginRequest:
      type: object
      description: Data required for logging in via email/password.
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          writeOnly: true
      required:
        - email
        - password
    LoginResponse:
      type: object
      description: Response containing the JWT access token for API clients. For browser clients, a cookie is typically set instead.
      properties:
        accessToken:
          type: string
          description: JWT access token.
        tokenType:
          type: string
          default: "Bearer"
          description: Type of the token (always Bearer).
      required:
        - accessToken
        - tokenType
...
    Error:
      type: object
      description: Standard error response format.
      properties:
        code:
          type: integer
          format: int32
          description: HTTP status code or application-specific code.
        message:
          type: string
          description: Detailed error message.
      required:
        - code
        - message

  # Reusable Responses
  responses:
    BadRequest:
      description: Invalid input (e.g., validation error, missing fields, invalid tag ID).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication failed (e.g., invalid credentials, invalid/expired token/cookie, missing authentication).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Conflict:
      description: Conflict (e.g., username or email already exists, tag name already exists for the user, resource state conflict).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

# Security Requirement applied globally or per-operation
# Most endpoints require either Bearer or Cookie auth.
security:
  - BearerAuth: []
  - CookieAuth: []

# API Path Definitions
paths:
...
  /auth/signup:
    post:
      summary: Register a new user via email/password (API).
      operationId: signupUserApi
      tags: [Auth]
      security: [] # No auth required to sign up
      requestBody:
        required: true
        description: User details for registration.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: User created successfully. Returns the new user object.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict" # e.g., Email or Username already exists
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/login:
    post:
      summary: Log in a user via email/password (API).
      description: Authenticates a user and returns a JWT access token in the response body for API clients. For browser clients, this endpoint typically also sets an HTTP-only cookie containing the JWT.
      operationId: loginUserApi
      tags: [Auth]
      security: [] # No auth required to log in
      requestBody:
        required: true
        description: User credentials for login.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful. Returns JWT token for API clients. Sets auth cookie for browsers.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
          headers:
            Set-Cookie: # Indicate that a cookie might be set for browser clients
              schema:
                type: string
              description: Contains the JWT authentication cookie (e.g., `jwt_token=...; HttpOnly; Secure; Path=/; SameSite=Lax`)
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized" # Invalid credentials
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /auth/google/login:
    get:
      summary: Initiate Google OAuth login flow.
      description: Redirects the user's browser to Google's authentication page. Not a typical REST endpoint, part of the web flow.
      operationId: initiateGoogleLogin
      tags: [Auth]
      security: [] # No API auth needed to start the flow
      responses:
        "302":
          description: Redirect to Google's OAuth consent screen. The 'Location' header contains the redirect URL.
          headers:
            Location:
              schema:
                type: string
                format: url
              description: URL to Google's OAuth endpoint.
        "500":
          description: Server error during redirect URL generation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/google/callback:
    get:
      summary: Callback endpoint for Google OAuth flow.
      description: Google redirects the user here after authentication. The server exchanges the received code for tokens, finds/creates the user, generates a JWT, sets the auth cookie, and redirects the user (e.g., to the web app dashboard).
      operationId: handleGoogleCallback
      tags: [Auth]
      security: [] # No API auth needed, Google provides auth code via query param
      # parameters:
      #   - name: code
      #     in: query
      #     required: true
      #     schema:
      #       type: string
      #     description: Authorization code provided by Google.
      #   - name: state
      #     in: query
      #     required: false # Recommended for security (CSRF protection)
      #     schema:
      #       type: string
      #     description: Opaque value used to maintain state between the request and callback.
      responses:
        "302":
          description: Authentication successful. Redirects the user to the frontend application (e.g., '/dashboard'). Sets auth cookie.
          headers:
            Location:
              schema:
                type: string
              description: Redirect URL within the application after successful login.
            Set-Cookie:
              schema:
                type: string
              description: Contains the JWT authentication cookie.
        "401":
          description: Authentication failed with Google or failed to process callback. Redirects to a login/error page.
          headers:
            Location:
              schema:
                type: string
              description: Redirect URL to an error or login page.
        "500":
          description: Internal server error during callback processing. Redirects to an error page.
          headers:
            Location:
              schema:
                type: string
              description: Redirect URL to an error page.

  # --- User Endpoints ---
...